Clean up and copy to getTraefikConfig
